Broke out the single massive file.
