+++ lib/impasta/helper.rb
@@ -1,0 +1,12 @@
+module Impasta
+  module ImpastaHelper
+    class << self
+      def method_missing_helper obj, meth
+        obj.impasta.methods << meth #[name, args, block]
+        yield
+      rescue ::NameError => error
+        raise MissingMethod.new self, error
+      end
+    end
+  end
+end
