+++ uspec/infiltrate_spec.rb
@@ -1,23 +1,15 @@
 # frozen_string_literal: true
 
 require_relative "spec_helper"
-require_relative "../lib/impasta/impasta_helper"
+require_relative "../lib/impasta/spies/infiltrate"
 
-klass = Impasta2
+klass = Impasta
 
-spec "tracks passed in messsages" do
-  imp = klass.dummy "whatever"
-  imp.foo "fooarg1"
-  imp.bar(){ p 'bar' }
-
-  imp.impasta.methods.map{|name,_,_| name} == [:foo,:bar] || imp.impasta.methods
-end
-
-spec "dissallows methods unknown to the source object" do
+spec "disallows methods unknown to the source object" do
   imp = klass.infiltrate Array
   begin
     imp.nonexistant_method
-  rescue klass::NoSuchMethod
+  rescue klass::MissingMethod
     true # correct error type returned
   end
 end
@@ -26,7 +18,7 @@ spec "captures errors and stores the parent error" do
   imp = klass.infiltrate Array
   begin
     imp.nonexistant_method
-  rescue klass::NoSuchMethod => error
+  rescue klass::MissingMethod => error
     error.parent_exception.is_a? NameError || error # this is the original error, should you need it
   end
 end
@@ -35,7 +27,7 @@ spec "captures errors and stores where the Impasta was instantiated" do
   imp = klass.infiltrate Array
   begin
     imp.nonexistant_method
-  rescue klass::NoSuchMethod => error
+  rescue klass::MissingMethod => error
     error.definition.include?("#{__FILE__}:#{__LINE__ - 4}") || error.definition # this is where you defined the Impasta object
   end
 end
@@ -52,7 +44,7 @@ spec "captures errors and stores the list of methods (including args/block) acce
 
   begin
     imp.nonexistant_method
-  rescue klass::NoSuchMethod => error
+  rescue klass::MissingMethod => error
     error.accessed_methods == expected_methods || error.accessed_methods # this will be a list of any of the methods called on it
   end
 end
@@ -61,7 +53,7 @@ spec "captures errors and stores the expected object instance" do
   imp = klass.infiltrate Array
   begin
     imp.nonexistant_method
-  rescue klass::NoSuchMethod => error
+  rescue klass::MissingMethod => error
     error.object == [] || error.object # this is what Impasta thinks the object should be
   end
 end
@@ -70,7 +62,7 @@ spec "captures errors and stores the missing method name" do
   imp = klass.infiltrate Array
   begin
     imp.nonexistant_method
-  rescue klass::NoSuchMethod => error
+  rescue klass::MissingMethod => error
     error.method_name == :nonexistant_method || error.method_name # this is the method that wasn't found
   end
 end
